split item and sheet
